requirements.txt
anyio==4.2.0
asttokens==2.2.0
azure-core==1.29.6
azure-cosmos==4.5.1
backcall==0.2.0
certifi==2023.11.17
charset-normalizer==3.3.2
colorama==0.4.6
contourpy==1.0.6
cycler==0.11.0
debugpy==1.6.4
decorator==5.1.1
entrypoints==0.4
executing==1.2.0
fonttools==4.38.0
idna==3.6
ipykernel==6.17.1
ipython==8.7.0
jedi==0.18.2
jupyter_client==7.4.7
jupyter_core==5.1.0
kiwisolver==1.4.4
lxml==4.9.2
matplotlib==3.6.2
matplotlib-inline==0.1.6
nest-asyncio==1.5.6
numpy==1.23.5
packaging==21.3
pandas==1.5.2
parso==0.8.3
pickleshare==0.7.5
Pillow==9.3.0
platformdirs==2.5.4
prompt-toolkit==3.0.33
psutil==5.9.4
pure-eval==0.2.2
Pygments==2.13.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-pptx==0.6.21
pytz==2022.6
pywin32==305
PyYAML==6.0.1
pyzmq==24.0.1
requests==2.31.0
six==1.16.0
sniffio==1.3.0
stack-data==0.6.2
structlog==23.3.0
tornado==6.2
traitlets==5.6.0
typing_extensions==4.9.0
urllib3==2.1.0
wcwidth==0.2.5
XlsxWriter==3.1.2
